Simplify CompactionStack in seperate resources class
